requirements_gdrive_integration.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
httplib2>=0.19.0
requests>=2.25.0
